==> 02-Train-Fraud-RGCN.ipynb <==
# Install DGL dependency

Please use notebook kernel with pytorch already installed. Using `conda_pytorch_p38` or `conda_pytorch_p36` will work. 
Install DGL dependency after selecting the kernel.

%pip install -q dgl

import numpy as np
import pandas as pd
from sklearn.metrics import roc_auc_score

import pandas as pd
from fgnn.fraud_detector import FraudRGCN

# Load train and test splits

df_train = pd.read_parquet('./data/train.parquet')

df_test = pd.read_parquet('./data/test.parquet')

# Set model parameters overloading defaults

We set parameters to match model parameters used in this [blog post](https://aws.amazon.com/blogs/machine-learning/build-a-gnn-based-real-time-fraud-detection-solution-using-amazon-sagemaker-amazon-neptune-and-the-deep-graph-library/).

params = {
    'embedding_size': 64,
    'n_layers': 2,
    'n_epochs': 150,
    'n_hidden': 16,
    'dropout': 0.2,
    'weight_decay': 5e-05,
    'lr': 0.01,
}

### print default model parameters
FraudRGCN()._default_params

# Train model in transductive mode
We train model five times and save trained models in `model/` directory. Note that to train the model in transductive mode, we combine train and test transactions and pass combined DataFrame to the `train_fg` method. We also pass `test_mask` that identifies test transactions with `True` values. In transductive mode, test labels will be masked out during training, but test samples will be used to construct nodes and features.

Training five models will take over half hour to complete.

import warnings
### disable CUDA-related warnings from torch library 
warnings.filterwarnings("ignore", category=UserWarning)

test_mask = [False]*len(df_train) + [True]*len(df_test)
for ii in range(1,6):
    fd = FraudRGCN()
    fd.train_fg(pd.concat([df_train, df_test], ignore_index=True), params=params, test_mask=test_mask)
    fd.save_fg(f"model/transductive_{ii}")

# Train model in inductive mode

We train model five times and save them to `model/` directory. Note that only training transactions are passed to the model for training. 

Training five models will take over half hour to complete.

for ii in range(1,6):
    fd = FraudRGCN()
    fd.train_fg(df_train, params=params)
    fd.save_fg(f"model/inductive_{ii}")
